+++ server/src/main/java/com/capstone/server/model/SearchHistoryEntity.java
@@ -46,7 +46,7 @@ public class SearchHistoryEntity {
     private Step step;
 
     @Builder.Default
-    private Integer searchRadius = 1; // 1km
+    private Double searchRadius = 1.0; // 1km
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "missing_people_id")
